Imported Upstream version 2.0.0+gitb2952297

+++ mspec/lib/mspec/helpers/tmp.rb
@@ -7,15 +7,20 @@ SPEC_TEMP_DIR = "#{File.expand_path(Dir.pwd)}/rubyspec_temp"
 
 SPEC_TEMP_UNIQUIFIER = "0"
 
+SPEC_TMEM_DIR_PID = Process.pid
+
 at_exit do
   begin
-    Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+    if SPEC_TMEM_DIR_PID == Process.pid
+      Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+    end
   rescue SystemCallError
     STDERR.puts <<-EOM
 
 -----------------------------------------------------
 The rubyspec temp directory is not empty. Ensure that
-all specs are cleaning up temporary files.
+all specs are cleaning up temporary files:
+  #{SPEC_TEMP_DIR}
 -----------------------------------------------------
 
     EOM
